requirements.txt
git+git://github.com/SiteSupport/gevent.git
glock
Routes
WebOb
docopt
storm
PyYAML
